# More control benchmarks and ScrollBar constructor optimization