## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Wanted to add more benchmarks for common controls so we can get better overview about the state of performance.

And while doing that I've noticed that `ScrollBar` is calculating visibility in a sub-optimal way (observables, bindings, LINQ - root of all the evil).

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
|        Method |     Mean |     Error |    StdDev |   Gen 0 |   Gen 1 |  Gen 2 | Allocated |
|-------------- |---------:|----------:|----------:|--------:|--------:|-------:|----------:|
| CreateTextBox | 1.010 ms | 0.0078 ms | 0.0069 ms | 39.0625 | 19.5313 | 1.9531 | 160.16 KB |

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
|        Method |     Mean |    Error |   StdDev |   Gen 0 |   Gen 1 |  Gen 2 | Allocated |
|-------------- |---------:|---------:|---------:|--------:|--------:|-------:|----------:|
| CreateTextBox | 949.9 us | 2.510 us | 2.225 us | 37.1094 | 17.5781 | 1.9531 | 152.58 KB |

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->
Use already existing `OnPropertyChanged<T>` method to check for changes. Got like 5% perf and memory improvement for free.

## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation